Proposed cheat-sheet changes from Quora question
I've read your quora question and made a few suggested changes for you to review and either accept or deny. I've also formatted your document as markdown which should display really nicely on github! You were on the right tracks and I'll explain in detail my changes in a comment and on quora for the purpose of archiving.